requirements.txt
flask==3.0.3
feedparser==6.0.11
newspaper3k==0.2.8
sumy==0.11.0
textblob==0.18.0.post0
scikit-learn==1.5.1
